+++ pkg/webui/console/api/index.js
@@ -72,13 +72,11 @@ export default {
     },
   },
   users: {
-    async get(userId) {
-      return instance.get(`${isBaseUrl}/users/${userId}`, {
-        headers: {
-          Authorization: `Bearer ${(await token()).access_token}`,
-        },
-      })
-    },
+    get: ttnClient.Users.getById.bind(ttnClient.Users),
+    list: ttnClient.Users.getAll.bind(ttnClient.Users),
+    update: ttnClient.Users.updateById.bind(ttnClient.Users),
+    delete: ttnClient.Users.deleteById.bind(ttnClient.Users),
+    search: ttnClient.Users.search.bind(ttnClient.Users),
     async authInfo() {
       return instance.get(`${isBaseUrl}/auth_info`, {
         headers: {