+++ pkg/webui/console/containers/header/index.js
@@ -28,6 +28,7 @@ import {
   mayViewApplications,
   mayViewGateways,
   mayViewOrganizationsOfUser,
+  mayManageUsers,
 } from '../../lib/feature-checks'
 
 @withRouter
@@ -40,6 +41,7 @@ import {
         mayViewApplications: checkFromState(mayViewApplications, state),
         mayViewGateways: checkFromState(mayViewGateways, state),
         mayViewOrganizations: checkFromState(mayViewOrganizationsOfUser, state),
+        mayManageUsers: checkFromState(mayManageUsers, state),
       }
     }
     return { user }
@@ -49,21 +51,34 @@ import {
 @bind
 class Header extends Component {
   static propTypes = {
+    /** Flag identifying whether links should be rendered as plain anchor link */
+    anchored: PropTypes.bool,
+    /** A handler for when the user clicks the logout button */
+    handleLogout: PropTypes.func.isRequired,
+    /** A handler for when the user used the search input */
+    handleSearchRequest: PropTypes.func,
+    mayManageUsers: PropTypes.bool,
+    mayViewApplications: PropTypes.bool,
+    mayViewGateways: PropTypes.bool,
+    mayViewOrganizations: PropTypes.bool,
+    /** A flag identifying whether the header should display the search input */
+    searchable: PropTypes.bool,
     /**
      * The User object, retrieved from the API. If it is `undefined`, then the
      * guest header is rendered
      */
-    anchored: PropTypes.bool,
-    /** Flag identifying whether links should be rendered as plain anchor link */
-    handleLogout: PropTypes.func.isRequired,
-    /** A handler for when the user used the search input */
-    handleSearchRequest: PropTypes.func,
-    /** A handler for when the user clicks the logout button */
-    searchable: PropTypes.bool,
-    /** A flag identifying whether the header should display the search input */
-    user: PropTypes.object,
-    /** The rights of the current user */
-    rights: PropTypes.rights,
+    user: PropTypes.user,
+  }
+
+  static defaultProps = {
+    anchored: false,
+    handleSearchRequest: () => null,
+    searchable: false,
+    user: undefined,
+    mayManageUsers: false,
+    mayViewApplications: false,
+    mayViewGateways: false,
+    mayViewOrganizations: false,
   }
 
   render() {
@@ -76,6 +91,7 @@ class Header extends Component {
       mayViewApplications,
       mayViewGateways,
       mayViewOrganizations,
+      mayManageUsers,
     } = this.props
 
     const navigationEntries = [
@@ -106,6 +122,12 @@ class Header extends Component {
     ]
 
     const dropdownItems = [
+      {
+        title: sharedMessages.userManagement,
+        icon: 'user_management',
+        path: '/admin/user-management',
+        hidden: !mayManageUsers,
+      },
       {
         title: sharedMessages.logout,
         icon: 'power_settings_new',